Move warnings to _core
