+++ parcels/_core/particleset.py
@@ -12,7 +12,7 @@ from zarr.storage import DirectoryStore
 from parcels._core.utils.time import TimeInterval, maybe_convert_python_timedelta_to_numpy
 from parcels._logger import logger
 from parcels._reprs import particleset_repr
-from parcels._warnings import ParticleSetWarning
+from parcels._core.warnings import ParticleSetWarning
 from parcels._core.kernel import Kernel
 from parcels.kernels import AdvectionRK4
 from parcels._core.particle import KernelParticle, Particle, create_particle_data
